More README.

+++ test/com/natpryce/pearlfish/example/calculator/Operands.java
@@ -1,9 +1,9 @@
 package com.natpryce.pearlfish.example.calculator;
 
 public class Operands {
-    public final int x, y;
+    public final long x, y;
 
-    public Operands(int x, int y) {
+    public Operands(long x, long y) {
         this.x = x;
         this.y = y;
     }
